Move Administration Table components to its own files
